Bugfixes for Physgun
Physgun was not properly deleting the game objects it was instantiated.
- [the LineRenderer was being destroyed but not the gameObject holding it]
this was fixed so it destroys the entire gameObject

- [the locationTarget was initializing twice, once creating a New GameObject() and then Initializing a copy of that new GameObject, and was only deleting the second one.]
changed this to be an additional empty gameObject held in the Physgun at all times, instead of taking extra processing to instantiate and delete it every time.

- [freezing something in place with the PhysGun would only freeze position]
this was fixed to FreezeAll