modified SamplePosition
adjusted SamplePosition to be based on the scale of Jeff, since he was having issues realigning to the NavMesh when too large or small.
making him smaller than .2f roughly still causes issues with the NavAgent, not sure how to solve though.

Additionally, added a timeout length to the Recovery(), so that if Jeff takes too long to realign to the NavMesh and Re-enable, he just respawns.